requirements.txt
awesome-slugify
beautifulsoup4
certifi
Django==4.2
django-crispy-forms
django-easy-maps
django-environ
django-filter
django-geoposition
django-location-field
django-olwidget
django-profiles
django-registration==3.3
django-rest-swagger==2.2.0
django-simple-captcha
django-social-auth==0.7.28
django-tables2==2.5.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
# drf-spectacular==0.26.2
drf-yasg==1.21.5
Faker
geopy
lxml
mysqlclient
Pillow
pylint-django
pymysql
python-memcached
requests

